// src/app/modules/hcm/shared/models/incidents/incidents-detail.ts
export class IncidentsDetail {
    idIncidents: number = -1;
    idEmploymentRelationship: number = -1;
    numEmployees: string = '';
    employeeFirstName: string = '';
    employeeSecondName: string = '';
    employeeLastName: string = '';
    employeeSecondLastName: string = '';
    idConcepts: number = -1;
    codConcepts: string = '';
    concepts: string = '';
    idPayrollType: number = -1;
    abbreviation: string = '';
    payrollType: string = '';
    idCalendar: number = -1;
    idStatus: number = -1;
    status: string = '';
    idUnit: number = -1;
    unit: string = '';
    valueIncident: number = -1;
    dateIncident: string = '';
    dateCreate: string = '';
    dateUpdate: string = '';
    selected: boolean = false;
}